+++ src/utilities/testing/index.ts
@@ -13,4 +13,5 @@ export { createMockClient } from './mocking/mockClient';
 export { stripSymbols } from './stripSymbols';
 export { default as subscribeAndCount } from './subscribeAndCount';
 export { itAsync } from './itAsync';
-export { withErrorSpy } from './withErrorSpy';
+export * from './withConsoleSpy';
+export * from './stringifyForDisplay';